protocol: Wrap up response handling
Hopefully this is more robust than before, and deduplicates some code.
The premise is that we should always receive at least 4 bytes (OK or
Err). Depending on whether the response is OK or Err, we can then wait
for the remainder of the response.

Hopefully this fixes lock-ups in the case of error.